pyproject.toml
[project]
name = "microalpha"
version = "0.1.0"
description = "A event-driven backtesting engine for quantitative strategies."
authors = [
    { name = "Mateo Bodon", email = "mateo.bodon@yale.edu" },
]
requires-python = ">=3.9"
dependencies = ["numpy", "pandas", "pyyaml"]

[project.optional-dependencies]
dev = [
    "pytest",
    "hypothesis",
    "mypy",
    "ruff",
    "black",
    "pytest-cov",
    "matplotlib",
    "pyarrow",
    "pydantic",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
exclude = [
    "notebooks/**",
    ".ipynb_checkpoints/**",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    # Temporarily ignore long lines in CI to unblock; keep Black for wrapping
    "E501",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
microalpha = "microalpha.cli:main"
